Move SiP unit tests
