+++ src/platform/forms/tests/save-in-progress/FormStartControls.unit.spec.jsx
@@ -5,8 +5,8 @@ import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { getFormDOM } from '../../../../src/platform/testing/unit/schemaform-utils';
-import { FormStartControls } from '../../../../src/applications/common/schemaform/save-in-progress/FormStartControls';
+import { getFormDOM } from '../../../testing/unit/schemaform-utils';
+import { FormStartControls } from '../../save-in-progress/FormStartControls';
 
 describe('Schemaform <FormStartControls>', () => {
   const startPage = 'testing';
