+++ src/platform/forms/tests/save-in-progress/RoutedSavableApp.unit.spec.jsx
@@ -3,8 +3,8 @@ import { expect } from 'chai';
 import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
 
-import { RoutedSavableApp } from '../../../../src/applications/common/schemaform/save-in-progress/RoutedSavableApp';
-import { LOAD_STATUSES, PREFILL_STATUSES } from '../../../../src/applications/common/schemaform/save-in-progress/actions';
+import { RoutedSavableApp } from '../../save-in-progress/RoutedSavableApp';
+import { LOAD_STATUSES, PREFILL_STATUSES } from '../../save-in-progress/actions';
 
 let oldWindow;
