+++ src/platform/forms/tests/save-in-progress/RoutedSavablePage.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
 
-import { RoutedSavablePage } from '../../../../src/applications/common/schemaform/save-in-progress/RoutedSavablePage';
+import { RoutedSavablePage } from '../../save-in-progress/RoutedSavablePage';
 
 describe('Schemaform <RoutedSavablePage>', () => {
   const location = {
