+++ src/platform/forms/tests/save-in-progress/RoutedSavableReviewPage.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
 
-import { RoutedSavableReviewPage } from '../../../../src/applications/common/schemaform/save-in-progress/RoutedSavableReviewPage';
+import { RoutedSavableReviewPage } from '../../save-in-progress/RoutedSavableReviewPage';
 
 describe('Schemaform save in progress: RoutedSavableReviewPage', () => {
   const location = {
