+++ src/platform/forms/tests/save-in-progress/SaveStatus.unit.spec.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import SkinDeep from 'skin-deep';
 import { expect } from 'chai';
 
-import SaveStatus from '../../../../src/applications/common/schemaform/save-in-progress/SaveStatus.jsx';
-import { SAVE_STATUSES } from '../../../../src/applications/common/schemaform/save-in-progress/actions';
+import SaveStatus from '../../save-in-progress/SaveStatus.jsx';
+import { SAVE_STATUSES } from '../../save-in-progress/actions';
 
 describe('<SaveStatus>', () => {
   const props = {
