+++ src/platform/forms/tests/save-in-progress/actions.unit.spec.js
@@ -18,9 +18,9 @@ import {
   setPrefillComplete,
   setFetchFormPending,
   setStartOver,
-} from '../../../../src/applications/common/schemaform/save-in-progress/actions';
+} from '../../save-in-progress/actions';
 
-import { logOut } from '../../../../src/platform/user/authentication/actions';
+import { logOut } from '../../../user/authentication/actions';
 
 let oldFetch;
 let oldSessionStorage;
